Integrated new PR from Copilot and removed the unrequired utils functions.
